Fix the build
Currently the project can not be build due to changes in the parent pom
see https://github.com/eclipse-osgi-technology/maven-pom/issues/12
