+++ runtime/registrar/pom.xml
@@ -41,5 +41,10 @@
       <groupId>org.osgi</groupId>
       <artifactId>org.osgi.service.component</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.osgi</groupId>
+      <artifactId>org.osgi.namespace.implementation</artifactId>
+      <version>1.0.0</version>
+    </dependency>
   </dependencies>
 </project>
