+++ tests/pom.xml
@@ -82,7 +82,9 @@
       <version>2.4.0</version>
       <type>jar</type>
     </dependency>
-
-
+    <dependency>
+      <groupId>org.osgi</groupId>
+      <artifactId>org.osgi.test.junit5</artifactId>
+    </dependency>
   </dependencies>
 </project>
