+++ tests/src/test/java/org/eclipse/osgi/technology/webservices/integration/tests/binding/WSEchoService.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     Christoph Läubrich - initial API and implementation
  *******************************************************************************/
-package org.eclipse.osgi.technology.webservices.tests.binding;
+package org.eclipse.osgi.technology.webservices.integration.tests.binding;
 
 import java.net.URL;
