+++ tests/src/test/java/org/eclipse/osgi/technology/webservices/integration/tests/package-info.java
@@ -9,4 +9,4 @@
  *
  * Contributors: Christoph Läubrich - initial API and implementation
  *******************************************************************************/
-package org.eclipse.osgi.technology.webservices.tests;
+package org.eclipse.osgi.technology.webservices.integration.tests;
